Break Button component apart
Make it easier to work with. The single file has a lot of context held
within itself. Also breaks apart helper files for the variant and size
styles for `StyledButton`
